Use UTC timestamps when authenticating with Emarsys
If Emarsys() was used without specifying a timezone, then the
resulting datetime object was not timezone aware. A timestamp without
timezone information is interpreted as a timestamp in the local
timezone of the Emarsys server (GMT+1 or GMT+2 during DST).
https://dev.emarsys.com/v2/before-you-start/authentication#timezone-considerations

I don't know the history of the original commit below, but it's
conceivable that relying on the default caused issues in some edge
cases. Either way it is a good idea to be explicit about the timezone,
and since we don't specify a timezone in our code and have no good
reason to do so, we can always use UTC and remove the tzinfo_obj.

This differs from the original commit, but results in an equivalent
change from the point of view of machtfit/signup (via django-emarsys).

Originally implemented thus:
https://github.com/eisensheng/python-emarsys/commit/aaa1085f879b29f15ff860766b0ae839914453e7

https://machtfit.atlassian.net/browse/MAC-2448